Split into Classes and Added Sub ID
This is the first time I split the Python code into classes and really did any refactoring. I also changed the default values from being in the script itself to being in a text file, as well as the user ids and API Keys. Finally, today was the first day that I actually got to use the new DatumBox Subscription ID they gave me, which allots me 10,000 requests instead of 1,000!