Refactored File System And Added Tests
This refactoring of the file system was done mainly so a strict documentation pattern was defined. This must be adhered to in the future. Each file of data must have 3 lines describing what the data is and how/why it was collected. The README.md file will document each search.

Also, between this commit and the last one, many tests were conducted, as outlined in the Readme.