First Product Sentiment Analysis
This commit is the first time we are going to try to do the full product sentiment analysis with all the pieces in place. Right now, our only problem is that the sentiment analysis engine is not very accurate (about %50 right). We will try to gather the data and find a correlation anyway.